// src/interfaces/Colors.ts
interface Colors {
    black: string
    red: string
    green: string
    yellow: string
    blue: string
    magenta: string
    cyan: string
    lightgray: string
    default: string
    darkgray: string
    lightred: string
    lightgreen: string
    lightyellow: string
    lightblue: string
    lightmagenta: string
    lightcyan: string
    white: string
    reset: string
}

export = Colors